src/local-vectors/faiss/faiss-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
faiss-cpu==1.12.0
voyageai
